Check in current state
