+++ .gitignore
@@ -75,6 +75,9 @@ xcschememanagement.plist
 xcuserdata/
 DerivedData/
 
+# VSCode project
+.vscode
+
 # Jetbrains project
 .idea/
 cmake-build-*/
